INTEGRATE CLEANADOCOPTS INTO GETADOCSTRING
cleanAdocOpts ensures strings used in asciidoctor
calls use only certain asciidoctor options. For example, the program
write to a pipe to asciidoctor and reads from asciidoctor's stdout, so
allowing the user to set the in and out files for asciidoctor would be
uneccessary and harmful.

But the function is only used once, so it can be integrated into the
reading of parsing of the dictionary of asciidoctor options into the
string that is passed to the call.