RETAIN ATTACHMENTS BETWEEN INPUT AND OUTPUT
All attachments (aside from the body of the message, which of course is
handled differently) are added, in the order they were on the original
message, to the outgoing message.

Attachments on the outgoing message are JUST attachments. Using the
add_attachment() method from the python email library wraps each
attachment in an rfc message, which is ugly and confusing to deal with
in a terminal / textual email client, and doesn't seem to add any
benefit for any kind of client. Instead the legacy attach() method is
used (which is used within the add_attachment method anyway) to just add
attachments as their own files.