REFACTOR MSG CREATION TO HANDLE SIMPLE AND MIXED
Issue #3 said that messages with inline images were being wrapped in
rfc822 message objects. This was actually affecting ALL non-mixed
messages. Fixed: text/plain input will result in multipart/alternative
output.

Bug was that the function to make alternative messages returned a
MIMEPart so that it could be attached when handling mixed messages, but
setting a multipart/alternative MIMEPart as the content of a message
creates an rfc822 object to hold that part. Fix was just a refactor.

Closes #3.