IMPLEMENT 'ASCIIDOCTOR-OPTIONS-STRING' OPTION
asciidoctor-options-string is an option which passes its value straight
to any invocation of asciidoctor.

It has to be used with '=' syntax and quotation, i.e.
`--asciidoctor-options-string="-a source-highlighter=pygments"` because
otherwise argparse things that that the args in the string are
themselves args for multi-mark-maker. This will need to go in the docs.

closes #5.